nettoyage, organisation du code
