+++ index.html
@@ -16,6 +16,6 @@
         <button id="playButton">Jouer au jeu</button>
     </div>
     
-    <script src="script.js"></script>
+    <script type="module" src="./js/script.js"></script>
 </body>
 </html>
